started client, need to change mutexes to map of shared_ptrs, lot of stuff wrong with that

+++ ServerSocket.cpp
@@ -8,12 +8,12 @@
 
 ServerSocket::ServerSocket() : port(0), fd(-1) { setup(); }
 
-ServerSocket::ServerSocket(int port) : port(port), fd(-1) { setup(); }
+ServerSocket::ServerSocket(uint16_t port) : port(port), fd(-1) { setup(); }
 
 ServerSocket::~ServerSocket() { cleanup(); }
 
 void ServerSocket::cleanup() {
-  // if socket open close fd
+  // If socket open close fd
   if (fd != -1) {
     close(fd);
     fd = -1;
@@ -21,34 +21,34 @@ void ServerSocket::cleanup() {
 }
 
 void ServerSocket::setup() {
-  // create socket
+  // Create socket
   fd = socket(AF_INET, SOCK_STREAM, 0);
   if (fd == -1) {
     throw std::runtime_error("Failed to create socket");
   }
 
-  // set socket to reuse address
+  // Set socket to reuse address
   int opt = 1;
   if (setsockopt(fd, SOL_SOCKET, SO_REUSEADDR, &opt, sizeof(opt)) == -1) {
     cleanup();
     throw std::runtime_error("Failed to set socket options");
   }
 
-  // configure socket to accept from any address and set port
+  // Configure socket to accept from any address and set port
   sockaddr_in server_addr{};
   std::memset(&server_addr, 0, sizeof(server_addr));
   server_addr.sin_family = AF_INET;
   server_addr.sin_addr.s_addr = INADDR_ANY;
   server_addr.sin_port = htons(port);
 
-  // bind socket to port
+  // Bind socket to port
   if (bind(fd, reinterpret_cast<sockaddr *>(&server_addr),
            sizeof(server_addr)) == -1) {
     cleanup();
     throw std::runtime_error("Failed to bind socket");
   }
 
-  // get port from newly created socket
+  // Get port from newly created socket
   socklen_t addr_len = sizeof(server_addr);
   if (getsockname(fd, reinterpret_cast<sockaddr *>(&server_addr), &addr_len) ==
       -1) {
@@ -57,20 +57,20 @@ void ServerSocket::setup() {
   }
   port = ntohs(server_addr.sin_port);
 
-  // set socket to listen with queue of 30
+  // Set socket to listen with queue of 30
   if (listen(fd, 30) == -1) {
     cleanup();
     throw std::runtime_error("Failed to listen on socket");
   }
 }
 
-int ServerSocket::get_port() { return port; }
+uint16_t ServerSocket::get_port() { return port; }
 
 int ServerSocket::accept_client() {
   sockaddr_in client_addr{};
   socklen_t client_addrlen = sizeof(client_addr);
 
-  // accept one connection from client
+  // Accept one connection from client
   int client_fd =
       accept(fd, reinterpret_cast<sockaddr *>(&client_addr), &client_addrlen);
   if (client_fd == -1) {
